useCart, cart.ts 리팩토링 (계산 로직 분리, let 제거, 선언형 메서드 사용(reduce)
